requirements.txt
quart==0.18.4
pyrogram
tgcrypto
coloredlogs==15.0.1
pyyaml==6.0
lxml==4.9.2
aiofile==3.8.5
tinydb==4.7.1
